Plugin Directory: Merge the cooldown and release-block tests into one file.
Both suites exercised the same fixture — a published plugin with a staged
release in cooldown and an update_source row serving the previous
version — duplicating ~90 lines of setup and helpers. Update_Source_Hold_Test
now carries both groups of tests with the fixture defined once, the block
tests reusing the cooldown tests' get_row()/set_status() helpers, and
get_release() gains its array|false return type.
